LayoutTests/fast/events/keydown-keypress-preventDefault-expected.txt

This tests that preventing the default behavior for a keydown event will not prevent the keypress event from firing, and the text will be inserted.
This matches IE7 and Firefox.
 
key down
key press
key down
key press
key down
key press
key down
key press
